# Conflicts:
#	src/main/java/edu/miu/cs/cs544/Security/SecurityConfiguration.java
#	src/main/java/edu/miu/cs/cs544/domain/Offering.java
#	src/main/resources/application.properties
#	src/test/java/edu/miu/cs/cs544/block/BlockServiceTest.java

+++ src/main/java/edu/miu/cs/cs544/repository/CourseRepository.java
@@ -2,13 +2,11 @@ package edu.miu.cs.cs544.repository;
 
 import edu.miu.cs.cs544.domain.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
 import java.util.Optional;
 
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Long> {
+public interface CourseRepository extends JpaRepository<Course, Integer> {
 
     Optional<Course> findByName(String name);
     Course findByCode(String name);
